+++ src/chart_types/xy_chart/state/selectors/compute_axis_ticks_dimensions.ts
@@ -24,11 +24,17 @@ import { getChartThemeSelector } from '../../../../state/selectors/get_chart_the
 import { getSettingsSpecSelector } from '../../../../state/selectors/get_settings_specs';
 import { CanvasTextBBoxCalculator } from '../../../../utils/bbox/canvas_text_bbox_calculator';
 import { AxisId } from '../../../../utils/ids';
-import { computeAxisTicksDimensions, AxisTicksDimensions, isDuplicateAxis } from '../../utils/axis_utils';
+import {
+  computeAxisTicksDimensions,
+  AxisTicksDimensions,
+  isDuplicateAxis,
+  defaultTickFormatter,
+} from '../../utils/axis_utils';
 import { computeSeriesDomainsSelector } from './compute_series_domains';
 import { countBarsInClusterSelector } from './count_bars_in_cluster';
+import { getAxesStylesSelector } from './get_axis_styles';
 import { getBarPaddingsSelector } from './get_bar_paddings';
-import { getAxisSpecsSelector } from './get_specs';
+import { getAxisSpecsSelector, getSeriesSpecsSelector } from './get_specs';
 import { isHistogramModeEnabledSelector } from './is_histogram_mode_enabled';
 
 /** @internal */
@@ -41,6 +47,8 @@ export const computeAxisTicksDimensionsSelector = createCachedSelector(
     getSettingsSpecSelector,
     computeSeriesDomainsSelector,
     countBarsInClusterSelector,
+    getSeriesSpecsSelector,
+    getAxesStylesSelector,
   ],
   (
     barsPadding,
@@ -50,13 +58,16 @@ export const computeAxisTicksDimensionsSelector = createCachedSelector(
     settingsSpec,
     seriesDomainsAndData,
     totalBarsInCluster,
+    seriesSpecs,
+    axesStyles,
   ): Map<AxisId, AxisTicksDimensions> => {
     const { xDomain, yDomain } = seriesDomainsAndData;
-
+    const fallBackTickFormatter = seriesSpecs.find(({ tickFormat }) => tickFormat)?.tickFormat ?? defaultTickFormatter;
     const bboxCalculator = new CanvasTextBBoxCalculator();
     const axesTicksDimensions: Map<AxisId, AxisTicksDimensions> = new Map();
     axesSpecs.forEach((axisSpec) => {
       const { id } = axisSpec;
+      const axisStyle = axesStyles.get(id) ?? chartTheme.axes;
       const dimensions = computeAxisTicksDimensions(
         axisSpec,
         xDomain,
@@ -64,7 +75,8 @@ export const computeAxisTicksDimensionsSelector = createCachedSelector(
         totalBarsInCluster,
         bboxCalculator,
         settingsSpec.rotation,
-        chartTheme.axes,
+        axisStyle,
+        fallBackTickFormatter,
         barsPadding,
         isHistogramMode,
       );