+++ src/chart_types/xy_chart/state/selectors/compute_series_domains.ts
@@ -38,6 +38,9 @@ export const computeSeriesDomainsSelector = createCachedSelector(
       customYDomainsByGroupId,
       deselectedDataSeries,
       settingsSpec.xDomain,
+      settingsSpec.orderOrdinalBinsBy,
+      // @ts-ignore blind sort option for vislib
+      settingsSpec.enableVislibSeriesSort,
     );
     return domains;
   },